GUVNOR-3323: Remove WiresGridsDemoPerspective and WiresTreesPerspective from Business Central
