--- uberfire-extensions/uberfire-wires/uberfire-wires-core/uberfire-wires-core-trees/src/main/java/org/uberfire/ext/wires/core/trees/client/factories/categories/TreeNodesCategory.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2015 JBoss, by Red Hat, Inc
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.uberfire.ext.wires.core.trees.client.factories.categories;
-
-import org.uberfire.ext.wires.core.api.factories.categories.Category;
-
-public class TreeNodesCategory extends Category {
-
-    public static final TreeNodesCategory CATEGORY = new TreeNodesCategory();
-
-    private TreeNodesCategory() {
-        super("Tree Nodes");
-    }
-}
