+++ uberfire-extensions/uberfire-wires/uberfire-wires-webapp/src/main/java/org/uberfire/ext/wires/client/demo/trees/factory/ExampleTreeNode1Factory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.uberfire.ext.wires.core.trees.client.factories;
+package org.uberfire.ext.wires.client.demo.trees.factory;
 
 import javax.enterprise.context.ApplicationScoped;
 
@@ -23,8 +23,7 @@ import org.uberfire.ext.wires.core.api.factories.categories.Category;
 import org.uberfire.ext.wires.core.api.shapes.WiresBaseShape;
 import org.uberfire.ext.wires.core.client.factories.AbstractBaseFactory;
 import org.uberfire.ext.wires.core.client.util.ShapesUtils;
-import org.uberfire.ext.wires.core.trees.client.factories.categories.TreeNodesCategory;
-import org.uberfire.ext.wires.core.trees.client.shapes.WiresExampleTreeNode1;
+import org.uberfire.ext.wires.client.demo.trees.shapes.WiresExampleTreeNode1;
 
 @ApplicationScoped
 public class ExampleTreeNode1Factory extends AbstractBaseFactory<Circle> {
