+++ uberfire-extensions/uberfire-wires/uberfire-wires-webapp/src/main/java/org/uberfire/ext/wires/client/demo/trees/palette/WiresTreesPaletteScreen.ui.xml
