+++ uberfire-extensions/uberfire-wires/uberfire-wires-webapp/src/main/java/org/uberfire/ext/wires/client/perspectives/WiresGridsDemoPerspective.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.uberfire.ext.wires.core.grids.client.perspectives;
+package org.uberfire.ext.wires.client.perspectives;
 
 import javax.enterprise.context.ApplicationScoped;
