+++ uberfire-extensions/uberfire-wires/uberfire-wires-webapp/src/main/resources/org/uberfire/ext/wires/client/resources/i18n/Constants_fr_FR.properties
