+++ uberfire-extensions/uberfire-wires/uberfire-wires-webapp/src/main/resources/org/uberfire/ext/wires/client/resources/i18n/Constants_zh_CN.properties
